✨ Added Resolution Scales ✨
One bug the resolution once changed goes to full screen. Attempted to add Controls, dont know where some of the scripts are being called. Leaving for later
