+++ Darkspire/Assets/GD/My Assets/Prefabs/UIPrefabs/SettingsMenu.prefab
@@ -632,7 +632,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 181907628566190880}
   m_Direction: 2
   m_Value: 0
-  m_Size: 1
+  m_Size: 0.75924665
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -1835,7 +1835,7 @@ RectTransform:
   m_Father: {fileID: 4609940669160557212}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 0.75924665}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 20}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -2082,7 +2082,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: -7.7, y: 0.0000121593475}
+  m_AnchoredPosition: {x: -7.7, y: 0.000011444092}
   m_SizeDelta: {x: -8.9248, y: 36.113}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!1 &5303170128415611274
@@ -3033,7 +3033,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &8504323858238432066
 RectTransform:
   m_ObjectHideFlags: 0
@@ -4040,9 +4040,9 @@ RectTransform:
   m_Father: {fileID: 8504323858238432066}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 0, y: 0}
+  m_SizeDelta: {x: -22.3588, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &607219450832629096
 MonoBehaviour:
