+++ Darkspire/Assets/GD/My Assets/Scripts/Settings/FullscreenOnorOff.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FullscreenOnorOff : MonoBehaviour
+{
+    public void SetFullscreen(bool isFullscreen)
+    {
+        Screen.fullScreen = isFullscreen;
+    }
+}
