+++ Darkspire/Assets/GD/My Assets/Scripts/Settings/FullscreenOnorOff.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a5e5025acee20e4893619e16ded34d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
