+++ Darkspire/Assets/GD/My Assets/Scripts/Settings/InGameSettings.cs
@@ -1,0 +1,64 @@
+using System.Collections.Generic;
+using TMPro;
+using UnityEngine;
+
+public class InGameSettings : MonoBehaviour
+{
+    #region ResolutionFields
+
+    [SerializeField] private TMP_Dropdown resolutionDropdown;
+
+    private Resolution[] resolutions;
+    private List<Resolution> filteredResolutions;
+
+    private float currentRefreshRate;
+    private int currentResolutionIndex = 0;
+
+    #endregion
+
+    private void Start()
+    {
+        resolutions = Screen.resolutions;
+        filteredResolutions = new List<Resolution>();
+
+        resolutionDropdown.ClearOptions();
+        currentRefreshRate = Screen.currentResolution.refreshRate;
+
+        for (int i = 0; i < resolutions.Length; i++)
+        {
+            if (resolutions[i].refreshRate == currentRefreshRate)
+            {
+                filteredResolutions.Add(resolutions[i]);
+            }
+        }
+
+        List<string> options = new List<string>();
+        for (int i = 0; i < filteredResolutions.Count; i++)
+        {
+            string resolutionOptions = filteredResolutions[i].width + "x" + filteredResolutions[i].height + " " + filteredResolutions[i].refreshRate + " Hz";
+            options.Add(resolutionOptions);
+            if (filteredResolutions[i].width == Screen.width && filteredResolutions[i].height == Screen.height)
+            {
+                currentResolutionIndex = i;
+            }
+        }
+
+        resolutionDropdown.AddOptions(options);
+        resolutionDropdown.value = currentResolutionIndex;
+        resolutionDropdown.RefreshShownValue();
+    }
+
+    public void ChangeResolution(int resolution)
+    {
+        if (resolution >= 0 && resolution < filteredResolutions.Count)
+        {
+            Resolution resolutionSet = filteredResolutions[resolution];
+            Screen.SetResolution(resolutionSet.width, resolutionSet.height, Screen.fullScreen);
+        }
+    }
+
+    public void SetFullscreen(bool isFullscreen)
+    {
+        Screen.fullScreen = isFullscreen;
+    }
+}
