+++ Darkspire/Assets/GD/My Assets/Scripts/Settings/InGameSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9485c271f15951f4db8c8f686897fdfc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
