--- Darkspire/Assets/GD/My Assets/Scripts/UI/MainMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3d38ff7c39ce0c54fa6f13181f2415d2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
