--- Darkspire/Assets/GD/My Assets/Scripts/UI/PauseMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 080974c96508b4049a9edb70c6f83dbb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
