--- Darkspire/Assets/GD/My Assets/Scripts/UI/Setting.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-public class Setting : MonoBehaviour
-{
-    public void SetFullscreen(bool isFullscreen)
-    {
-        Screen.fullScreen = isFullscreen;
-     
-    }
-}
