+++ Darkspire/UserSettings/EditorUserSettings.asset
@@ -15,10 +15,10 @@ EditorUserSettings:
       value: 555556565c500d595f0c0e73447508444316492e757d7e32792f4830bbb66d3e
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 55000654000d510e5a59557642775e44414e1c7a7e7e71687e711f35b0e6673e
+      value: 5502015452505f0d555b5c2648720744444f487b7529723678701e62b1b3643e
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5502015452505f0d555b5c2648720744444f487b7529723678701e62b1b3643e
+      value: 55000654000d510e5a59557642775e44414e1c7a7e7e71687e711f35b0e6673e
       flags: 0
     UnityEditor.ShaderGraph.Blackboard:
       value: 18135939215a0a5004000b0e15254b524c030a3f2964643d120d1230e9e93a3fd6e826abbd3c3e302a07a37e0901373ae01e0008f707250d171df81a53a541475f9148954a9a4b489a4d4c989f489da1d8e9fce5cae2fdd7d0cba1fedbf6e8ebf0e5e7a1fae6ac8ce8a2c182a6d8c2869bd1f5dc8e9c9b97ea9c819ffcca90898d86d598899ac5
